Rename AletheiaInput and use basic input in it
